Moved test data/schema for test_feature_helper. Updated test_feature_helper data and schema import.
