+++ projects/_03_independent_cqc/_04_feature_engineering/tests/utils/test_feature_helper.py
@@ -2,12 +2,18 @@ import unittest
 import warnings
 
 from utils.utils import get_spark
+
 from projects._03_independent_cqc._04_feature_engineering.utils import helper as job
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    ModelFeatures as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
+    ModelFeatures as Schemas,
+)
+
 from utils.column_names.ind_cqc_pipeline_columns import (
     IndCqcColumns as IndCQC,
 )
-from tests.test_file_data import ModelFeatures as Data
-from tests.test_file_schemas import ModelFeatures as Schemas
 
 
 class LocationsFeatureEngineeringTests(unittest.TestCase):
